+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.0.1
 
 * 2024-02-02 - Initial commit
+* 2024-02-12 - Added Haptic feedback, animation duration, and scale factor 0.0 to 1.0